Restructure includes of FloatingPointComparator
