+++ include/deal.II/matrix_free/dof_info.h
@@ -21,7 +21,6 @@
 #include <deal.II/base/config.h>
 
 #include <deal.II/base/exceptions.h>
-#include <deal.II/base/floating_point_comparator.h>
 #include <deal.II/base/vectorization.h>
 
 #include <deal.II/matrix_free/face_info.h>
@@ -46,6 +45,9 @@ namespace internal
     class HangingNodes;
 
     struct TaskInfo;
+
+    template <typename Number>
+    struct ConstraintValues;
   } // namespace MatrixFreeFunctions
 } // namespace internal
 
@@ -79,39 +81,6 @@ namespace internal
      * is in hanging_nodes_internal.h.
      */
     using compressed_constraint_kind = std::uint8_t;
-
-    /**
-     * A struct that takes entries describing a constraint and puts them into
-     * a sorted list where duplicates are filtered out
-     */
-    template <typename Number>
-    struct ConstraintValues
-    {
-      ConstraintValues();
-
-      /**
-       * This function inserts some constrained entries to the collection of
-       * all values. It stores the (reordered) numbering of the dofs
-       * (according to the ordering that matches with the function) in
-       * new_indices, and returns the storage position the double array for
-       * access later on.
-       */
-      template <typename number2>
-      unsigned short
-      insert_entries(
-        const std::vector<std::pair<types::global_dof_index, number2>>
-          &entries);
-
-      std::vector<std::pair<types::global_dof_index, double>>
-                                           constraint_entries;
-      std::vector<types::global_dof_index> constraint_indices;
-
-      std::pair<std::vector<Number>, types::global_dof_index> next_constraint;
-      std::map<std::vector<Number>,
-               types::global_dof_index,
-               FloatingPointComparator<VectorizedArray<Number>>>
-        constraints;
-    };
 
     /**
      * The class that stores the indices of the degrees of freedom for all the
