+++ include/deal.II/matrix_free/matrix_free.templates.h
@@ -40,6 +40,7 @@
 
 #include <deal.II/lac/dynamic_sparsity_pattern.h>
 
+#include <deal.II/matrix_free/constraint_info.h>
 #include <deal.II/matrix_free/face_info.h>
 #include <deal.II/matrix_free/face_setup_internal.h>
 #include <deal.II/matrix_free/hanging_nodes_internal.h>
